Added Affinity Scores For all Precompute Models
This commit adds the affinity score, a number between 0.01 and 1,000 which
represents how strongly the model recommends the given item. We are
calculating the affinity score as (number of purchases | number of views, etc)
normalized from 0.01 to 1,000 with a lower-bound for the normalization assumed
to be 0. We use 'ceil' rather than 'round' to ensure we don't drop down to 0.00
and the minimum actually remains 0.01.

https://kibo.atlassian.net/browse/DEC3-262
https://kibo.atlassian.net/browse/DEC3-207